LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/path-objects/2d.path.isPointInPath.basic.2-expected.txt

2d.path.isPointInPath.basic.2

isPointInPath() detects whether the point is inside the path


PASS isPointInPath() detects whether the point is inside the path 
